+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Make resources' (last resort) Close with Finalizers opt-in with GuardWithFinalizers().
+
 ## [v0.39.2]
 ### Changed
 - simplify handleDeadline, hopefully eliminiate SIGSEGVs